refactors pv, moves msg to consume.rs
